Eduardo

+++ Dto/FilterCommons/PaginationParams.cs
@@ -2,6 +2,6 @@ namespace AicaDocsApi.Dto.FilterCommons;
 
 public class PaginationParams
 {
-    public short PageNumber { get; set; } = 1;
-    public short PageSize { get; set; } = 10;
+    public int PageNumber { get; set; } = 1;
+    public int PageSize { get; set; } = 10;
 }
